Moved Monuments related utils into their own class
