+++ app/src/test/kotlin/fr/free/nrw/commons/UtilsTest.kt
@@ -1,6 +1,6 @@
 package fr.free.nrw.commons
 
-import fr.free.nrw.commons.utils.Utils
+import fr.free.nrw.commons.utils.Monuments
 import org.junit.Test
 import org.junit.jupiter.api.Assertions
 import java.util.Calendar
@@ -10,20 +10,20 @@ class UtilsTest {
     fun wikiLovesMonumentsYearBeforeSeptember() {
         val cal = Calendar.getInstance()
         cal.set(2022, Calendar.FEBRUARY, 1)
-        Assertions.assertEquals(2021, Utils.getWikiLovesMonumentsYear(cal))
+        Assertions.assertEquals(2021, Monuments.getWikiLovesMonumentsYear(cal))
     }
 
     @Test
     fun wikiLovesMonumentsYearInSeptember() {
         val cal = Calendar.getInstance()
         cal.set(2022, Calendar.SEPTEMBER, 1)
-        Assertions.assertEquals(2022, Utils.getWikiLovesMonumentsYear(cal))
+        Assertions.assertEquals(2022, Monuments.getWikiLovesMonumentsYear(cal))
     }
 
     @Test
     fun wikiLovesMonumentsYearAfterSeptember() {
         val cal = Calendar.getInstance()
         cal.set(2022, Calendar.DECEMBER, 1)
-        Assertions.assertEquals(2022, Utils.getWikiLovesMonumentsYear(cal))
+        Assertions.assertEquals(2022, Monuments.getWikiLovesMonumentsYear(cal))
     }
 }
